@@ -14,7 +14,7 @@
     <name>search</name>
     <description>Using Java API client</description>
     <properties>
-        <java.version>17</java.version>
+        <java.version>19</java.version>
         <springdoc-openapi-ui.version>1.6.15</springdoc-openapi-ui.version>
         <jackson-databind-nullable.version>0.2.6</jackson-databind-nullable.version>
         <swagger-annotations.version>1.6.8</swagger-annotations.version>
@@ -31,18 +31,14 @@
         <openapi-spec.file>${project.basedir}/src/main/resources/api.yml</openapi-spec.file>
         <openapi-spec.classpath>classpath:api.yml</openapi-spec.classpath>
         <api.path>com.elasticsearch.search</api.path>
-        <maven.compiler.source>17</maven.compiler.source>
-        <maven.compiler.target>17</maven.compiler.target>
+        <maven.compiler.source>19</maven.compiler.source>
+        <maven.compiler.target>19</maven.compiler.target>
     </properties>
     <dependencies>
         <!-- Spring Boot -->
         <dependency>
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-web</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-starter-thymeleaf</artifactId>
         </dependency>
         <dependency>
             <groupId>org.springframework.boot</groupId>
